More warning fixes

+++ src/umpire/strategy/MixedPool.cpp
@@ -71,7 +71,7 @@ void* MixedPool::allocate(size_t bytes)
 
   void* mem;
 
-  if (index < m_fixed_pool.size()) {
+  if (index < static_cast<int>(m_fixed_pool.size())) {
     // allocate in fixed pool
     mem = m_fixed_pool[index].allocate();
     m_map[reinterpret_cast<uintptr_t>(mem)] = index;
